refactor: TopBar 분리
