+++ app/src/main/java/nextstep/shoppingcart/ui/cart/ShoppingCartScreen.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import nextstep.shoppingcart.data.Cart
 import nextstep.shoppingcart.R
+import nextstep.shoppingcart.data.Cart
 import nextstep.shoppingcart.ui.component.ShoppingCartButton
 import nextstep.shoppingcart.ui.component.ShoppingCartNavigationTopBar
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
