+++ app/src/main/java/nextstep/shoppingcart/ui/component/ShoppingCartNavigationTopBar.kt
@@ -2,7 +2,6 @@ package nextstep.shoppingcart.ui.component
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.ShoppingCart
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -10,6 +9,7 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
 import nextstep.shoppingcart.R
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -33,23 +33,12 @@ fun ShoppingCartNavigationTopBar(
     )
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
+@Preview
 @Composable
-fun ShoppingCartActionsTopBar(
-    title: String,
-    onActionClick: () -> Unit
-) {
-    CenterAlignedTopAppBar(
-        title = {
-            Text(text = title)
-        },
-        actions = {
-            IconButton(onClick = onActionClick) {
-                Icon(
-                    imageVector = Icons.Filled.ShoppingCart,
-                    contentDescription = stringResource(id = R.string.content_description_shopping_cart)
-                )
-            }
-        }
+private fun ShoppingCartNavigationTopBarPreview() {
+    ShoppingCartNavigationTopBar(
+        title = "상품 상세",
+        onNavigationClick = { },
     )
 }
+
